No Quotes on Header Fields
The current decision is that header fields should
not be quoted.  This may seem a bit strange if
people are looking at the header as being like
an OBJECT!, but if it's seen as being like a dialect
then having it look nicer makes sense...given that
it's not evaluated.