Make USCII Spec File Much More Dialected
The original USCII code did not make much use of the dialecting aspect
of Rebol, instead just demo'ing basic features like having images built in,
and being able to put documentation and binary literals into the file.
However, this misses much of what makes the language unique.

This reimagines the specification to use the parts more creatively, such
as putting the abbreviation as a SET-WORD! and the character codepoint
number inside a GROUP!.  So instead of looking like a standard JSON
object, the characters are identified by something that serves as a nice
visual divider and is compacted into one line.

Then the character data itself is represented not as strings but as 7
sequential WORD!s that are made up of on-and-off Unicode boxes.

At the moment this uses a simple UPARSE rule to transform the data
into the old format, so it can run through the same Redbol code that
was working before.