Add .gitattributes to disable syntax highlight
The syntax highlighting of Rebol files works against the "no
keywords" mentality of the language.  It would be nice if there
was a version that just highlighted lexical elements (such as
making SET-WORD!s bold) but the version that GitHub runs cannot
be configured, and so it colors "keywords" which is ill-suited
to dialects.